# backend/docs/swagger.yaml
basePath: /
definitions:
...
  providers.ProviderDownloadResponse:
    properties:
      arch:
        type: string
      download_url:
        type: string
      filename:
        type: string
      os:
        type: string
      protocols:
        items:
          type: string
        type: array
      shasum:
        type: string
      shasums_signature_url:
        type: string
      shasums_url:
        type: string
      signing_keys:
        $ref: '#/definitions/providers.ProviderSigningKeys'
    type: object
  providers.ProviderGPGKey:
    properties:
      ascii_armor:
        type: string
      key_id:
        type: string
    type: object
...
  providers.ProviderSearchItem:
    properties:
      description:
        type: string
      download_count:
        type: integer
      id:
        type: string
      latest_version:
        type: string
      namespace:
        type: string
      source:
        type: string
      type:
        type: string
    type: object
  providers.ProviderSearchMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  providers.ProviderSearchResponse:
    properties:
      meta:
        $ref: '#/definitions/providers.ProviderSearchMeta'
      providers:
        items:
          $ref: '#/definitions/providers.ProviderSearchItem'
        type: array
    type: object
  providers.ProviderSigningKeys:
    properties:
      gpg_public_keys:
        items:
          $ref: '#/definitions/providers.ProviderGPGKey'
        type: array
    type: object
...
info:
  contact:
    email: admin@your-registry.example.com
    name: Registry Support
  description: Complete Terraform Module and Provider Registry with SCM integration,
    mirrors, and storage backends
  license:
    name: Apache-2.0
  termsOfService: https://your-registry.example.com/terms
  title: Terraform Registry API
  version: 1.0.0
paths:
...
  /api/v1/providers/search:
    get:
      description: Search for providers by name or namespace with pagination.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by namespace
        in: query
        name: namespace
        type: string
      - description: Maximum results to return (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/providers.ProviderSearchResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Search providers
      tags:
      - Providers
...
  /v1/providers/{namespace}/{type}/{version}/download/{os}/{arch}:
    get:
      description: Returns download URL, checksums, and signing key info for a specific
        provider platform. Implements the Terraform Provider Registry Protocol.
      parameters:
      - description: Provider namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Provider type (e.g. aws, azurerm)
        in: path
        name: type
        required: true
        type: string
      - description: Semantic version (e.g. 1.2.3)
        in: path
        name: version
        required: true
        type: string
      - description: Target OS (e.g. linux, darwin, windows)
        in: path
        name: os
        required: true
        type: string
      - description: Target architecture (e.g. amd64, arm64)
        in: path
        name: arch
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/providers.ProviderDownloadResponse'
        "400":
          description: Invalid version or platform
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Provider, version, or platform not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Download provider platform binary
      tags:
      - Providers
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"JWT token or API key. For JWT: ''Bearer {token}''. For API Key:
      ''Bearer {api_key}''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"